requirements.txt
astroid==1.5.3
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
decorator==4.0.11
idna==2.5
ipython==6.1.0
ipython-genutils==0.2.0
isort==4.2.15
jedi==0.10.2
lazy-object-proxy==1.3.1
mccabe==0.6.1
motor==1.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.2
pyasn1==0.2.3
pycparser==2.18
Pygments==2.2.0
PyJWT==1.5.2
pylint==1.7.2
pymongo==3.4.0
pytz==2017.2
pyuv==1.3.0
requests==2.18.1
simplegeneric==0.8.1
six==1.10.0
tornado==4.5.1
tornaduv==0.4.0
traitlets==4.3.2
urllib3==1.21.1
uvloop==0.8.0
wcwidth==0.1.7
wrapt==1.10.10
